Run forge fmt

+++ script/CNC.s.sol
@@ -48,7 +48,7 @@ contract CNCContractScript is Script, Test {
     uint32 t3OreTokenId = 6;
     uint32 shipLumberConsidered = 30_000;
     uint32 shipOreConsidered = 15_000;
-    uint256 wraithEthConsidered = 0.050 ether;
+    uint256 wraithEthConsidered = 0.05 ether;
     uint256 clockworkEthConsidered = 0.015 ether;
     uint32 resourcesOffered = 10_000;
 
@@ -59,18 +59,12 @@ contract CNCContractScript is Script, Test {
     uint32 maxCertificateCampaignRedemptions = 6_500; // for the lootbox/certs campaign
     uint32 maxCampaignRedemptions = 100_000_000; // for all other campaigns
 
-    string certCampainURI =
-        "ipfs://QmfEGHX8SNSavpVQ1kJUJ2wKx2gJqprAwRdw2BKFbbqK1v";
+    string certCampainURI = "ipfs://QmfEGHX8SNSavpVQ1kJUJ2wKx2gJqprAwRdw2BKFbbqK1v";
 
-    function testLootboxRedeem(
-        address lootboxesAdddr,
-        address certificatesAddr
-    ) public {
-        uint campaignId = 1;
-        uint tokenId = 1;
-        ERC721ShipyardRedeemableMintable lootboxes = ERC721ShipyardRedeemableMintable(
-                lootboxesAdddr
-            );
+    function testLootboxRedeem(address lootboxesAdddr, address certificatesAddr) public {
+        uint256 campaignId = 1;
+        uint256 tokenId = 1;
+        ERC721ShipyardRedeemableMintable lootboxes = ERC721ShipyardRedeemableMintable(lootboxesAdddr);
 
         Certificates certificates = Certificates(certificatesAddr);
 
@@ -85,14 +79,7 @@ contract CNCContractScript is Script, Test {
         assertEq(certificates.balanceOf(msg.sender, 6), 0);
 
         uint256[] memory traitRedemptionTokenIds = new uint256[](0);
-        bytes memory data = abi.encode(
-            campaignId,
-            0,
-            bytes32(0),
-            traitRedemptionTokenIds,
-            uint256(0),
-            bytes("")
-        );
+        bytes memory data = abi.encode(campaignId, 0, bytes32(0), traitRedemptionTokenIds, uint256(0), bytes(""));
 
         uint256[] memory tokenIds = new uint256[](1);
         tokenIds[0] = tokenId;
@@ -111,10 +98,7 @@ contract CNCContractScript is Script, Test {
         assertEq(lootboxes.balanceOf(msg.sender), 0);
     }
 
-    function setUpCertificatesCampaign(
-        address lootboxesAddr,
-        address certificatesAddr
-    ) public returns (uint256) {
+    function setUpCertificatesCampaign(address lootboxesAddr, address certificatesAddr) public returns (uint256) {
         // Setups the certificates campaign.
         // A single lootbox should yield a certificate.
 
@@ -182,12 +166,9 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint campaignId = certificates.createCampaign(campaign, certCampainURI);
+        uint256 campaignId = certificates.createCampaign(campaign, certCampainURI);
         // uint campaignId = 1;
         // certificates.updateCampaign(
         //     campaignId,
@@ -197,14 +178,11 @@ contract CNCContractScript is Script, Test {
         return campaignId;
     }
 
-    function setUpWraithBlueprintCampaign(
-        address shipsAddr,
-        address certificatesAddr,
-        address resourcesAddr
-    ) public returns (uint256) {
-        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(
-                shipsAddr
-            );
+    function setUpWraithBlueprintCampaign(address shipsAddr, address certificatesAddr, address resourcesAddr)
+        public
+        returns (uint256)
+    {
+        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(shipsAddr);
 
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
@@ -277,28 +255,22 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint campaignId = ships.createCampaign(campaign, "");
+        uint256 campaignId = ships.createCampaign(campaign, "");
 
         return campaignId;
     }
 
-    function setUpWraithGoldprintCampaign(
-        address shipsAddr,
-        address certificatesAddr,
-        address resourcesAddr
-    ) public returns (uint256) {
+    function setUpWraithGoldprintCampaign(address shipsAddr, address certificatesAddr, address resourcesAddr)
+        public
+        returns (uint256)
+    {
         // Setups the goldprint campaign for the wraith ship
         // Considerations:
         // - certificate with the goldprint dynamic trait
         // - offer: a ship
-        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(
-                shipsAddr
-            );
+        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(shipsAddr);
 
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
@@ -344,29 +316,20 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint campaignId = ships.createCampaign(
-            campaign,
-            "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP"
-        );
+        uint256 campaignId = ships.createCampaign(campaign, "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP");
 
         return campaignId;
     }
 
-    function setUpClockworkBlueprintCampaign(
-        address shipsAddr,
-        address certificatesAddr,
-        address resourcesAddr
-    ) public returns (uint256) {
+    function setUpClockworkBlueprintCampaign(address shipsAddr, address certificatesAddr, address resourcesAddr)
+        public
+        returns (uint256)
+    {
         // Creates a campaigns for clockwork ship from a blueprint.
 
-        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(
-                shipsAddr
-            );
+        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(shipsAddr);
 
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
@@ -439,30 +402,21 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint campaignId = ships.createCampaign(
-            campaign,
-            "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP"
-        );
+        uint256 campaignId = ships.createCampaign(campaign, "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP");
         return campaignId;
     }
 
-    function setUpClockworkGoldprintCampaign(
-        address shipsAddr,
-        address certificatesAddr,
-        address resourcesAddr
-    ) public returns (uint256) {
+    function setUpClockworkGoldprintCampaign(address shipsAddr, address certificatesAddr, address resourcesAddr)
+        public
+        returns (uint256)
+    {
         // Setups the goldprint campaign
         // Considerations:
         // - certificate with the goldprint dynamic trait
         // - offer: a ship
-        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(
-                shipsAddr
-            );
+        ERC721ShipyardRedeemableMintable ships = ERC721ShipyardRedeemableMintable(shipsAddr);
 
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
@@ -507,48 +461,22 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint campaignId = ships.createCampaign(
-            campaign,
-            "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP"
-        );
+        uint256 campaignId = ships.createCampaign(campaign, "ipfs://QmQjubc6guHReNW5Es5ZrgDtJRwXk2Aia7BkVoLJGaCRqP");
 
         return campaignId;
     }
 
-    function setUpShipCampaigns(
-        address shipsAddr,
-        address certificatesAddr,
-        address resourcesAddr
-    ) public {
+    function setUpShipCampaigns(address shipsAddr, address certificatesAddr, address resourcesAddr) public {
         // Sets up all four ship campaigns
-        setUpWraithBlueprintCampaign(
-            shipsAddr,
-            certificatesAddr,
-            resourcesAddr
-        );
+        setUpWraithBlueprintCampaign(shipsAddr, certificatesAddr, resourcesAddr);
 
-        setUpWraithGoldprintCampaign(
-            shipsAddr,
-            certificatesAddr,
-            resourcesAddr
-        );
+        setUpWraithGoldprintCampaign(shipsAddr, certificatesAddr, resourcesAddr);
 
-        setUpClockworkBlueprintCampaign(
-            shipsAddr,
-            certificatesAddr,
-            resourcesAddr
-        );
+        setUpClockworkBlueprintCampaign(shipsAddr, certificatesAddr, resourcesAddr);
 
-        setUpClockworkGoldprintCampaign(
-            shipsAddr,
-            certificatesAddr,
-            resourcesAddr
-        );
+        setUpClockworkGoldprintCampaign(shipsAddr, certificatesAddr, resourcesAddr);
     }
 
     function setUpResourcesCampaign(
@@ -603,66 +531,32 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint256 campaignId = Resources(resourcesAddr).createCampaign(
-            campaign,
-            ""
-        );
+        uint256 campaignId = Resources(resourcesAddr).createCampaign(campaign, "");
         return campaignId;
     }
 
-    function setUpResourcesCampaigns(
-        address resourcesAddr,
-        address certificatesAddr
-    ) public {
+    function setUpResourcesCampaigns(address resourcesAddr, address certificatesAddr) public {
         // Sets up all the resources campaigns
         setUpResourcesCampaign( // t1 lumber
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT1Lumber,
-            t1LumberTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT1Lumber, t1LumberTokenId);
         setUpResourcesCampaign( // t2 lumber
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT2Lumber,
-            t2LumberTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT2Lumber, t2LumberTokenId);
         setUpResourcesCampaign( // t3 lumber
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT3Lumber,
-            t3LumberTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT3Lumber, t3LumberTokenId);
         setUpResourcesCampaign( // t1 ore
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT1Ore,
-            t1OreTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT1Ore, t1OreTokenId);
         setUpResourcesCampaign( // t2 ore
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT2Ore,
-            t2OreTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT2Ore, t2OreTokenId);
         setUpResourcesCampaign( // t3 ore
-            certificatesAddr,
-            resourcesAddr,
-            traitValueT3Ore,
-            t3OreTokenId
-        );
+        certificatesAddr, resourcesAddr, traitValueT3Ore, t3OreTokenId);
     }
 
-    function setUpCosmeticsCampaign(
-        address certificatesAddr,
-        address cosmeticsAddr,
-        bytes32 traitValue
-    ) public returns (uint256) {
+    function setUpCosmeticsCampaign(address certificatesAddr, address cosmeticsAddr, bytes32 traitValue)
+        public
+        returns (uint256)
+    {
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
             itemType: ItemType.ERC721_WITH_CRITERIA,
@@ -706,53 +600,22 @@ contract CNCContractScript is Script, Test {
             manager: msg.sender
         });
 
-        Campaign memory campaign = Campaign({
-            params: params,
-            requirements: requirements
-        });
+        Campaign memory campaign = Campaign({params: params, requirements: requirements});
 
-        uint256 campaignId = ERC721ShipyardRedeemableMintable(cosmeticsAddr)
-            .createCampaign(campaign, "uri://");
+        uint256 campaignId = ERC721ShipyardRedeemableMintable(cosmeticsAddr).createCampaign(campaign, "uri://");
         return campaignId;
     }
 
-    function setUpCosmeticsCampaigns(
-        address cosmeticsAddr,
-        address certificatesAddr
-    ) public {
+    function setUpCosmeticsCampaigns(address cosmeticsAddr, address certificatesAddr) public {
         setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueHat1);
         setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueHat2);
         setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueHat3);
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueEyepiece1
-        );
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueEyepiece2
-        );
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueEyepiece3
-        );
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueCannon1
-        );
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueCannon2
-        );
-        setUpCosmeticsCampaign(
-            certificatesAddr,
-            cosmeticsAddr,
-            traitValueCannon3
-        );
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueEyepiece1);
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueEyepiece2);
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueEyepiece3);
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueCannon1);
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueCannon2);
+        setUpCosmeticsCampaign(certificatesAddr, cosmeticsAddr, traitValueCannon3);
     }
 
     function mintAndSetTraits(address certificatesAddr) public {
@@ -770,10 +633,10 @@ contract CNCContractScript is Script, Test {
         // address receiver = msg.sender;
         address receiver = 0xb9131A703595189c021014d85F74e7BFCb30f8D6;
 
-        for (uint i = 1; i < 2; i++) {
-            uint tokenIdBase = 400000 + (i * 100);
-            for (uint j = 1; j <= 19; j++) {
-                uint tokenId = tokenIdBase + j;
+        for (uint256 i = 1; i < 2; i++) {
+            uint256 tokenIdBase = 400000 + (i * 100);
+            for (uint256 j = 1; j <= 19; j++) {
+                uint256 tokenId = tokenIdBase + j;
                 bytes32 traitValue = bytes32(j);
 
                 certificates.mint(receiver, tokenId, 1); // certificate
